[config] Refactor dynamic-config
The goal is to have a common library (config-builder) to generate
configs but to have 3 different entry points to simplify the utilities
API.

This is the precursor before implementating the fullnode-config-builder

Closes: #1963
Approved by: davidiw
