+++ docker/validator-dynamic/docker-run-dynamic.sh
@@ -29,6 +29,9 @@ if [ -n "${CFG_SEED_PEER_IP}" ]; then # Seed peer ip for discovery
 	    params+="--bootstrap /ip4/${CFG_SEED_PEER_IP}/tcp/6180 "
 fi
 
-/opt/libra/bin/dynamic-config-builder --data-dir /opt/libra/etc --output-dir /opt/libra/etc/ ${params[@]}
+/opt/libra/bin/validator-config-builder \
+    --data-dir /opt/libra/etc \
+    --output-dir /opt/libra/etc/ \
+    ${params[@]}
 
 exec /opt/libra/bin/libra-node -f /opt/libra/etc/node.config.toml
