+++ docker/validator-dynamic/validator-dynamic.Dockerfile
@@ -21,8 +21,8 @@ COPY . /libra
 RUN cargo build --release -p libra-node -p client -p dynamic-config-builder && cd target/release && rm -r build deps incremental
 
 ### Production Image ###
-FROM libra_e2e:latest AS validator_with_config
-COPY --from=config_builder /libra/target/release/dynamic-config-builder /opt/libra/bin
+FROM libra_e2e:latest as validator_with_config
+COPY --from=config_builder /libra/target/release/validator-config-builder /opt/libra/bin
 COPY docker/validator-dynamic/docker-run-dynamic.sh /
 CMD /docker-run-dynamic.sh
