+++ testsuite/cluster-test/src/cluster.rs
@@ -5,7 +5,7 @@
 
 use crate::{aws::Aws, instance::Instance};
 use anyhow::{ensure, format_err, Result};
-use dynamic_config_builder::init_from_seed;
+use dynamic_config_builder::ValidatorConfig;
 use generate_keypair::load_key_from_file;
 use libra_config::config::AdmissionControlConfig;
 use libra_crypto::ed25519::{Ed25519PrivateKey, Ed25519PublicKey};
@@ -120,7 +120,7 @@ impl Cluster {
         let seed = "1337133713371337133713371337133713371337133713371337133713371337";
         let seed = hex::decode(seed).expect("Invalid hex in seed.");
         let seed = seed[..32].try_into().expect("Invalid seed");
-        let (mint_key, _) = init_from_seed(seed);
+        let (_, mint_key) = ValidatorConfig::new().seed(seed).build_faucet_client().expect("Unable to build faucet keys");
         let mint_key_pair = KeyPair::from(mint_key);
         Ok(Self {
             instances,
